Update OpenWrt version in vars
The OpenWrt version is used to point to the package repository. The used
used version ist defined by the commit hash in `gluon/modules`. Also see
in `openwrt/include/version.mk` of the specified hash in the OpenWrt
repository.

If not specified corectly the `opkg` command will fail to fetch
packages not provided by ffp.